SOAPCommunication/src/test/java/dk/sdu/sem4/pro/soap/SOAPCommunicationTest.java
package dk.sdu.sem4.pro.soap;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.BeforeEach;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.testng.annotations.Test;

import javax.xml.soap.*;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertEquals;

class SOAPCommunicationTest {

    @Mock
    private SOAPConnectionFactory soapConnectionFactory;

    @Mock
    private SOAPConnection soapConnection;

    @Mock
    private SOAPMessage soapMessage;

    @BeforeEach
    void init() throws SOAPException {
        MockitoAnnotations.openMocks(this);
        when(soapConnectionFactory.createConnection()).thenReturn(soapConnection);
        when(soapConnection.call(any(SOAPMessage.class), any())).thenReturn(soapMessage);
    }

    @Test
    void testReceive() throws Exception {
        // Assume `createSOAPRequest` and `processSOAPResponse` are properly implemented
        SOAPCommunication communicator = new SOAPCommunication(soapConnectionFactory);
        Document result = communicator.receive();

        // Your assertions go here
        assertNotNull(result);
        // More detailed assertions can be added based on the expected structure of your SOAP response
    }

    @Test
    void testSend() throws Exception {
        // Assume `createSOAPRequest` and `processSOAPResponse` are properly implemented
        SOAPCommunication communicator = new SOAPCommunication(soapConnectionFactory);
        Integer result = communicator.send(null); // Replace 'null' with a proper Document instance if needed

        // Your assertions go here
        assertNotNull(result);
        // More detailed assertions can be added based on the expected structure of your SOAP response
    }
}
